requirements.txt
babel==2.9.0
Flask==2.0.3
Jinja2==3.1.1
python-dateutil==2.6.0
flask-moment==0.11.0
flask-wtf==0.14.3
flask_sqlalchemy==2.4.4
virtualenv
SQLAlchemy
Flask-Migrate
json
logging
forms
datetime
